Fixed nssdb.add_cert() for HSM.
The nssdb.add_cert() has been modified to import certificates
properly. If HSM is used, the certificate will be imported into
HSM without trust attributes. If trust attributes are specified,
the certificate will be imported into internal token as well with
the trust attributes. If no HSM is used, the certificate will be
imported into the internal token with the trust attributes if
available.

https://pagure.io/dogtagpki/issue/2901

Change-Id: I4027b3064694ecf41bc616cf1b67581e4d103531