Set nextUpdate in OCSP responses
Some OCSP clients adhere to the Lightweight OCSP Profile (RFC 5019)
which requires that the OCSP response include the nextUpdate field.

Update the CA subsystem's OCSP responder to include the nextUpdate
field when it is configured to use the CRL cache.  The nextUpdate
field in the OCSP response is set to the nextUpdate time of the
"master" CRL issuing point.

If the OCSP responder is not configured to use the CRL cache, there
is no reasonable value for nextUpdate.  In this case, we continue to
omit it.

Fixes: https://pagure.io/dogtagpki/issue/2661
Change-Id: Idbf7354b0ecc45c0498c4b7c05458f726f40336f