Fix Bug 1501436 - TPS CS.cfg should be reflected with the changes after an in-place upgrade.
This upgrade script will add the needed config params to an existing CS.cfg for TPS.

The params consist of the params required for the token profile : externalRegISEtoken.

The code also grabs the unsecure phone home url out of the instances's server.xml.
This way the new profile is configured exactly like what happens when doing a pkispawn.
The correct nonsecure url will be in place.

Added some review changes. Also we modified the python properties file class to be able to
handle a property value that happens to contain the delimeter "=". Ex name=cn=people.

Change-Id: I2478013b396082ffdc3d99ed86a821ec86ac4c5d