parseGeneralName: properly parse iPAddress GN with netmask
There are a couple of problems with iPAddress general name parsing
(primarily used for the Name Constraints extension).

First, an IP address with netmask expressed as e.g.
1.2.3.4,255.0.0.0 or ::1,ffff:: is outright rejected, causing
issuance failure with a message like:

  NameConstraintsExtDefault: createExtension
    netscape.security.x509.InvalidIPAddressException: Invalid IP
    Address '10.10.10.10,255.255.255.0'

Second, an IPv4 address with CIDR-style netmask is misinterpreted as
an IPv6 address _without_ netmask, e.g. the input "192.168.1.1/24"
gets misinterpreted as "c0a8:1c8:ffff:ffff:000:000:000:000", which
is not a conforming value in the Name Constraints extension.

To resolve these problems, separate the handling of these two cases
and fix the logic.  A new class, CIDRNetmask, does the heavy lifting
in the CIDR netmask case.

Consider the following configuration (irrelevant keys and key
prefixes omitted for brevity). It contains values which caused
failures or incorrect outputs:

  nameConstraintsExcludedSubtreeNameChoice_0=IPAddress
  nameConstraintsExcludedSubtreeNameValue_0=10.10.10.10/24
  nameConstraintsExcludedSubtreeNameChoice_1=IPAddress
  nameConstraintsExcludedSubtreeNameValue_1=10.10.10.10,255.255.255.0
  nameConstraintsExcludedSubtreeNameChoice_2=IPAddress
  nameConstraintsExcludedSubtreeNameValue_2=dead:beef::1/128
  nameConstraintsExcludedSubtreeNameChoice_3=IPAddress
  nameConstraintsExcludedSubtreeNameValue_3=dead:beef::,ffff:ffff::

This configuration now succeeds and produces the correct output.
The extension value produced using the above configuration is (per
OpenSSL pretty print):

  X509v3 Name Constraints: critical
    Excluded:
      IP:10.10.10.10/255.255.255.0
      IP:10.10.10.10/255.255.255.0
      IP:DEAD:BEEF:0:0:0:0:0:1/FFFF:FFFF:FFFF:FFFF:FFFF:FFFF:FFFF:FFFF
      IP:DEAD:BEEF:0:0:0:0:0:0/FFFF:FFFF:0:0:0:0:0:0

Part of: https://pagure.io/dogtagpki/issue/2922
Change-Id: I61d5fcceadcca28cc951802ee4b95691653dd356