GeneralNameInterface: methods for checking name validity
Some general names may be valid only for describing a single subject
(e.g. Subject Alt Name extension), or for describing a range of
subjects (e.g. Name Constraints extension).  For example, an
iPAddress name MUST have 4 (IPv4) or 16 (IPv6) octets in the
"single" context, or 8 (IPv4) or 32 (IPv6) octets in range context.

Add the validSingle() and validSubtree() methods to
GeneralNameInterface and all implementing classes.  These methods
can be used to check whether the value is valid for use in the
corresponding context.

Part of: https://pagure.io/dogtagpki/issue/2922

Change-Id: Ib77286b309f1d505fe15313483ec658a55780f83