Fix lightweight CA key replication
The resolution for issue https://pagure.io/dogtagpki/issue/2654
caused a regression in lightweight CA key replication.  When the
authorityMonitor encounters a CA whose keys are not present,
signingUnit initialisation fails (as expected).  The signing info
event logging behaviour introduced in commit
4551eb1ce6b14e4a37f9c70b3bfd6c9050e13f10 then results in a
NullPointerException, crashing the authorityMonitor thread.

Fix the issue by extracting the signing info event logging behaviour
to a separate method, and invoke that method as the final step of
signingUnit initialisation.

Fixes: https://pagure.io/dogtagpki/issue/2929
Change-Id: Ic6663c09c30754f4fb914dcaf0bc2d902aa91473