PKIConnection.get: time out after 5s
There is a contention between the timeouts of PKIConnection.get (the
default for connect(2)) and Instance.wait_for_startup (60s).  When
/etc/hosts contains an IP address for the host which is routable but
not responded to (e.g. during FreeIPA installation with --setup-dns
and --ip-address=<not-yet-existant>), the connection attempt causes
pkispawn() to block for a long duration.  By the time it unblocks,
the Instance.wait_for_startup() timeout has been exceeded and no
further connection attempts are made.  Installation fails.

Avoid this situation by setting a timeout of 5 seconds on
PKIConnection.get().

Fixes: https://pagure.io/dogtagpki/issue/2939
Change-Id: Id746faee9bd9a2a61bbc15f55d9ccbc652997bf1