Update patch set 2
Patch Set 2: Code-Review+2

This looks good for now, but we probably need to reconsider where to put the migration script because of this scenario:

Suppose someone upgraded to PKI 10.6.1 on F27 in the future, the NSS migration will not be executed since F27 still uses dbm. Then if the system is later upgraded to F28, the NSS migration will not be executed either since the upgrade script was written for 10.6.0.

It might be best to run the migration when the server is restarted.

Patch-set: 2
Reviewer: Endi S. Dewata <1010221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2