Fixed pki-server cert CLI.
The pki-server cert CLI has been changed such that it can run
based on the list of certificates defined in CS.cfg even if the
certificate themselves are not available yet in the NSS database.

The security_databases.py has been modified to store the system
cert nicknames and token names in CS.cfg such that the pki-server
cert CLI can be used to export certificates during installation.

https://pagure.io/dogtagpki/issue/203

Change-Id: I046cfa7d45e2d0ae7b6de353d0840db0899789f7