Ticket #2940 [MAN] Missing Man pages for tools CMCRequest, CMCResponse, CMCSharedToken, and CMCRevoke
This patch adds man pages for CMCRequest, CMCResponse, and CMCSharedToken.
In addition, the usage in CMCResponse has been enhanced to include a
verbose mode which will output certs in Base64 encoding individually.
A "note" has been added to CMCRevoke --help to direct users to CMCRequest
for better usability. The man page for CMCRevoke is intentionaly left out
for this reason.

The URL in CMCRequest.1 is a placeholder for the follow-up patch.  It will
be replaced once the examples are complete.

This patch addresses https://pagure.io/dogtagpki/issue/2940

Change-Id: Id1df31a29207a0d12d50b7a3b959a3abcd9748d0