Update patch set 3
Patch Set 3:

Christina, I updated the commit message to include more info/examples about the cases that previously failed, and the extension values that they now produce (OpenSSL pretty-print).

The requested refactoring is performed in a new commit at the tip of the branch.

No changes to toString were performed.  We don't really need to produce CIDR notation there; IMO it is probably best not to, so the output more accurately reflects the datum that goes into the certificate.

Patch-set: 3