Updated compose scripts to create full source tarballs
Previously the compose scripts for core, console, and theme would
create different tarballs which contained the relevant sources
only even though they are coming from the same source repository.

To reduce maintenance, the compose scripts will now generate
identical tarballs that contain the complete source files in the
repository.

In this patch the tarballs will still have different names, but
in a subsequent patch they will be changed to use the same name.

https://pagure.io/dogtagpki/issue/2923

Change-Id: I453b044fa8ecb9df16b4f81a2aac942ed0f9fd55