Update patch set 1
Patch Set 1: Code-Review+2

Not sure if we should remove the default constructor. We might eventually create a REST API for ACL, and use the ACL class as a parameter, and usually it needs a default constructor. But it's easy to add it back if needed. So I'll ACK, but it's up to you if you want to clean it up.

Patch-set: 1
Reviewer: Endi S. Dewata <1010221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2