Update patch set 1
Patch Set 1:

Thanks for the review, Endi.  I'd rather delete the code that's not being used.  Definitely prefer to fully validate and initialise on construction anyway; construct then setX, setY, setZ is less safe and more burdensome on the programmer.

If there is an important reason to restore it later (e.g. for RESTEasy) then we can do so.

Patch-set: 1
Label: Code-Review=0