Update patch set 1
Patch Set 1:

Ah, I relased that the missing name attribute is probably why the CI is failing, so I modified this patch to include the fix.

The change is trivial, so I'll keep all your ACKs and merge when the CI has verified the changes.  Thanks for the reviews, Endi!

Patch-set: 1
Label: Code-Review=0