Update patch set 1
Patch Set 1:

I haven't looked at the code in details, but here's my initial comment.

To my understanding the only file that needs to be hidden and placed in the main folder is .travis.yml (because that's where Travis will look for it), but everything else does not need to follow that requirement.

So let's move all Travis files (the new ones being added here and the existing .travis_run_task.sh) into the .travis folder to reduce the clutter in the main folder, and remove the .travis_ prefix from the file names.

In another patch we can rename ".travis" to "travis" to make it visible and easier to access during development.

Patch-set: 1