Fix  #2996 ECC installation for non CA subsystems needs improvement.
The problem is that the installation of say a KRA, which is ECC enabled fails out of the box.

This is due to the fact that the internal cert profiles for the following certificates is incorrect:

1. sslserver cert
2. subsystem cert
3. admin cert

In the ECC case there is some hard coding that references the well known cert profiles for RSA versions of the above certs.

What we need in the ECC case is a way to correctly select the ECC versions of the above profiles.
Therefore this fix does the following:

1. Makes the selection of either the ECC version or the RSA version of the above internal cert profiles based on the key type, ecc or rsa. This solution relies upon well known profile names, but can be modified in the future to be more customizable , should the need arise.

2. I found a related problem when trying to create a ECC enabled KRA in a SHARED instance scenario. There was some final cloning related config code that was grossly RSA specific and throws exceptions when ECC is involved. I altered this piece of code to skip over the bad things with ECC and let the RSA case run unimpeded. We may need further refinement for the ECC case, but I felt this was needed to allow something like an ECC kra to be installed in a shared instance scenario.

Change-Id: I192dc18e50c87403624dd46754c5f22bc988d9a7