Fix banner file loading
The banner code was loading the banner file with
codecs.open(filename, 'UTF-8'), but the second argument to codecs.open()
is not an encoding but a mode.

Since Dogtag no longer supports Python 2.6, the io.open() function does a
much better job here. It's equivalent to Python 3's open() builtin. By
default, it loads text files with UTF-8 codec.

Change-Id: I2fbaea04bb313bdaf21ceaa0c0c68d0cfcd5ea9a
Signed-off-by: Christian Heimes <cheimes@redhat.com>