Handle empty NameConstraints subtrees when reading extension
When reading stored NameConstraints extension data on a request, if
includedSubtrees or excludedSubtrees is empty, an exception is
thrown.  But these are valid cases, so do not thrown an exception.

Also perform some minor drive-by refactors and add the 'static'
qualifier to a few methods to improve readability.

Part of: https://pagure.io/dogtagpki/issue/2922

Change-Id: I925d8a64b96dd0f45b0548ceb11dbee4223cd64c