Ticket #2959 Address pkispawn ECC profile overrides
This preliminary patch shows a quick design for an idea to address ticket 2959.

The idea is to use the % replacement strings as part of the profile names in
the default.cfg file for pkispawn,
and change the profile names to mach the format. So for example:

%(pki_admin_key_type)AdminCert.profile

would either be translated to rsaAdminCert.profile or eccAdminCert.profile
depending  on the value in pki_admin_key_type

This preliminary patch has been tested to work. This patch is not intended as
a full patch, and will be updated with more complete changes such as man page
changes.

fixes https://pagure.io/dogtagpki/issue/2959

Change-Id: I9a9f70e415438e0b4130294abb725c74fd6e1b95