Refactored CMake variables for theme
The BUILD_DOGTAG_PKI_THEME and BUILD_REDHAT_PKI_THEME variables
have been replaced with a single THEME variable. If not specifed,
it will default to "dogtag". If it's empty, the theme packages
will not be build. If it's not empty, the theme packages will be
built with the specified theme.

Change-Id: I913fa670a41795da61746c2acddac981c2f84a84