getTheSerialNumber: only return null if next range not available
When cloning, if the master's current number range has been depleted
due to a previous UpdateNumberRange request,
Repository.getTheSerialNumber() returns null because the next serial
number is out of the current range, but the next range has not been
activated yet.  NullPointerException ensues.

Update getTheSerialNumber() to return the next serial number even
when it exceeds the current number range, as long as there is a next
range.  If there is no next range, return null (as before).  It is
assumed that the next range is non-empty

Also do a couple of drive-by method extractions to improve
readability.

Part of: https://pagure.io/dogtagpki/issue/3055