Add missing synchronisation for range management
Several methods in Repository (and CertificateRepository) need
synchronisation on the intrisic lock.  Make these methods
synchronised.

Also take the lock in UpdateNumberRange so that no serial numbers
can be handed out in other threads between peekNextSerialNumber()
and set(Next)?MaxSerial().  Without this synchronisation, it is
possible that the master instance will use some of the serial
numbers it transfers to the clone.

Fixes: https://pagure.io/dogtagpki/issue/3055