Repository: handle depleted range in initCache()
Repository.initCache() does not handle the case where the current
range has been fully depleted, but the switch to the next range has
not occurred yet.  This situation arises when the range has been
fully depleted by servicing UpdateNumberRange requests for clones.

Detect this situation and handle it by switching to the next range
(when available).

Part of: https://pagure.io/dogtagpki/issue/3055