Check for missing FQDNs during pkispawn
When installing via pkispawn on a system with no hostname set,
or hostname not correctly set in /etc/hosts, raise an exception
early in the install process. This prevents deploys where the
certificates are assigned to localhost.localdomain; in this
scenario, creating a clone from this CA will fail as the clone
cannot validate the certificates of the CA master.

This adds a new command line option, --strict-hostname, to pkispawn
to enforce strict hostname checking.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>