Updated pki-server <subsystem>-audit-event-find
The pki-server <subsystem>-audit-event-find has been modified
to return only the audit events that are applicable to the
subsystem based on the information stored in the comments in
audit-events.properties.

The comments in audit-events.properties have been modified such
that they can be parsed more easily to get the list of audit
events and their applicable subsystems.

The information about the applicable subsystems will be added
in subsequent patches.

https://pagure.io/dogtagpki/issue/2686