cert-fix: sleep after starting server
If the server does not start quickly enough, cert-fix sends requests
to the server before it is ready to handle them, causing failure.

A proper solution is to poll the server until the status resource
indicates that it is ready.  But for now, the quick workaround is to
sleep for a little while.

Part of: https://pagure.io/dogtagpki/issue/2776