PKIInstance: add 'cert_folder' and 'cert_file' methods
The cert_folder and locations of certificates under that folder are
useful to know from outside the PKIInstance class.  In particular
the cert-fix tool will need these data.  Extract the computation of
the folder path to a property, and the computation of certificate
file paths to a method.

Part of: https://pagure.io/dogtagpki/issue/2776