cert-fix: use admin password authentication
If the agent/admin certificate is expired, cert-fix will fail.
Avoid this issue by using password authentication to submit the
renewal requests.

We don't know the current admin account password (and the user might
not know it either), so we have to reset it.  This will be a caveat
of cert-fix.  But because the user does know the Directory Manager
password, they can reset the admin account password afterwards.

Part of: https://pagure.io/dogtagpki/issue/2776