cert-fix: add CLI option to specify agent account
The name of the Dogtag admin account is configurable.  The current
hardcoded value, "admin", is correct for FreeIPA deployment but may
be incorrect for others.  In particular, the default admin account
name id "caadmin".  Furthermore, an operator may wish cert-fix to
use a particular agent or admin account.

Teach cert-fix the --agent-uid option which specifies the admin
account to use.

Part of: https://pagure.io/dogtagpki/issue/2776