cert-fix: use LDAPI
The LDAP password modify extended operation requires confidentiality
(i.e. TLS/STARTTLS).  If the LDAP service certificate is expired,
ldappasswd fails.

To avoid this problem, use LDAPI.  Teach cert-fix the --ldapi-socket
option, which gives the location of the LDAPI socket and which is
required.

This change introduces a new assumption, namely that LDAPI and
autobind are enable, and that the autobind user (typically root) is
mapped to an account with sufficient privileges (typically
"cn=Directory Manager".

Part of: https://pagure.io/dogtagpki/issue/2776