Allow page size to influence LDAP query size
When performing an LDAP query, we need to take into account the actual
page size of the incoming request. Otherwise, our LDAP query can either
overflow or underflow the request's page size. However, we can't blindly
set SIZELIMIT either; instead, treat it as a maximum value which we
can't exceed.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>