Updated PKI server library
The deployment scriptlet has been modified to link the server
library folder instead of creating a folder with links to
individual library files.

An upgrade script has been added to make the same changes in
existing instances.

The code that regenerates the links to individual library files
for Tomcat migration is no longer needed and has been removed.