Added pki-server run
A new pki-server run command has been added to run PKI server
in the foreground instead of in the background as systemd service.

By default the server will run with the same UID used by the
server's systemd service, but the command provides an option to
run the server as the current user, which is needed for run the
server in containers.

The command will also include the JAVA_OPTS specified in the
Tomcat configuration file (e.g. debugging parameters).