Don't raise an exception on missing of 'pkcs11.txt'
The lack of `pkcs11.txt` (if `key4.db` and `cert9.db` are present)
is the symptom of not completed upgrade of NSS db (dbm -> sql).

The partial upgrade happens on open of NSS db with NSS_INIT_NOMODDB.
In this case only certificates and keys are automatically migrated
to sql storage. Such mode is used by certmonger during initial
reading of the tracked NSS certificates.

This results in subsequent attempts to run PKI migration fail with:
"ERROR: /var/lib/pki/pki-tomcat/alias contains an incomplete NSS database in SQL format".

So, in this case migration should fallback to dbm NSS and perform upgrade
if the latter is present.

Fixes: https://pagure.io/dogtagpki/issue/3104
Signed-off-by: Stanislav Levin <slev@altlinux.org>