Fix noise generation for EC certificates
When generating noise for elliptic curves, very few bytes of entropy are
required (in comparison to RSA) because EC private keys are random data,
not random primes. Thus the amount of available entropy just need to be
sufficient for the size of the curve.

Rather than dealing with a mapping between curve to its size, set a
fixed value of 1024 bytes.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>