Fix parameters for EC-based CSR generation
When generating EC-based certificate requests, we incorrectly used
key_size as the -g parameter. This is correct for RSA keys, but
incorrect for EC keys (as the parameter is generally ignored).
Compounding to this, key_size (under key_type == 'ecc') is actually the
name of the curve, and not the size of the key under that curve.

We fix the parameter generation to support both the curve and the
key_size as the curve name.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>