Bug 1744095 - CMCResponse is not working as expected
This patch fixes the issue that HttpClient extracting less bytes than
the actual data size from the HTTP response.
My investigation shows that there used to be 6 lines of headers, and
now it's down to 5.
The fix is to default to 5, but add an unadvertised numHeaderLines
that allows one to customize in case the server changes again.
It is limited to the range of 1 - 56

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1744095