Updated certificate request review process
This patch introduces new certificate request review processes
which should be easier to use and automate.

The following command will display a summary of the request,
then ask the user to enter an action:

  $ pki ca-cert-request-review <request ID>

The following command will display a summary of the request,
then ask the user to confirm the specified action:

  $ pki ca-cert-request-<action> <request ID>

The following command will execute the specified action on
the request without asking for confirmation:

  $ pki ca-cert-request-<action> <request ID> --force

The following commands will store the complete request into
a file allowing a more detailed review, then perform the
specified action based on the updated request in the file:

  $ pki ca-cert-request-review <request ID> --output-file <file>
  $ pki ca-cert-request-<action> <request ID> --input-file <file>

The old processes are still available, but they have been
deprecated and may be removed in the future.

https://www.dogtagpki.org/wiki/PKI_10.8_PKI_CLI_Changes