Cleaned up TPS build scripts
Previously the TPS build scripts generated some artifacts in the
buildroot that were not included in the RPM package so rpmbuild
would generate warnings about those files.

To avoid the warnings the TPS build scripts have been modified
to no longer install those files into the buildroot.

In the future the unused sources should be removed from the
source repository.